Experiment aspect correlation with global prior
This commit add a new experiment on correlated aspects. The key difference is
to utilize global co-rating information of the same product category
instead of only focusing on the product co-rating (sparse number).

Even with global co-rating prior, highly correlated aspects normally get
correlation of only 0.5xxx. Therefore, the correlated uncertainty is
adapted for this fact. That is to possibly switch to other highly
correlated aspect's uncertainty if their correlation passes a threshold,
which is 0.5 by default.